Version 0.4.1 Update (First Version) FST
The build up to first release version went through five parts, the basics of which are included in Version Log, which will later be amended to main readme.

This update includes a playable build included with source, which I will soon test, to make sure the downloaded source can be assembled in Godot.

Monetization for this game and current release is donation based. Assets were selected based on their license, which is not included.